+++ src/album/entities/album.entity.ts
@@ -1,1 +1,6 @@
-export class Album {}
+export interface Album {
+  id: string;
+  name: string;
+  year: number;
+  artistId: string | null;
+}
