+++ src/favorites/entities/favorite.entity.ts
@@ -1,1 +1,5 @@
-export class Favorite {}
+export interface Favorites {
+  artists: string[];
+  albums: string[];
+  tracks: string[];
+}
