--- src/interfaces/album.interface.ts
@@ -1,6 +1,0 @@
-export interface Album {
-  id: string;
-  name: string;
-  year: number;
-  artistId: string | null;
-}
