--- src/interfaces/favorites.interface.ts
@@ -1,5 +1,0 @@
-export interface Favorites {
-  artists: string[];
-  albums: string[];
-  tracks: string[];
-}
