--- src/interfaces/track.interface.ts
@@ -1,7 +1,0 @@
-export interface Track {
-  id: string;
-  name: string;
-  artistId: string | null;
-  albumId: string | null;
-  duration: number;
-}
