--- src/interfaces/user.interface.ts
@@ -1,8 +1,0 @@
-export interface User {
-  id: string;
-  login: string;
-  password: string;
-  version: number;
-  createdAt: number;
-  updatedAt: number;
-}
