+++ src/track/entities/track.entity.ts
@@ -1,1 +1,7 @@
-export class Track {}
+export interface Track {
+  id: string;
+  name: string;
+  artistId: string | null;
+  albumId: string | null;
+  duration: number;
+}
