+++ src/user/entities/user.entity.ts
@@ -1,1 +1,8 @@
-export class User {}
+export interface User {
+  id: string;
+  login: string;
+  password: string;
+  version: number;
+  createdAt: number;
+  updatedAt: number;
+}
